LayoutTests/media/modern-media-controls/slider/slider-knob-none-expected.txt

Testing that Slider with KnobStyle.None has no visible gap.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS knobStyle.display is "none"
PASS primaryStyle.flexGrow is "50"
PASS trackStyle.flexGrow is "50"
PASS primaryRight is trackLeft
PASS successfullyParsed is true

TEST COMPLETE
